Separate the page tsx file and the index
